Availability spreadsheet
I chose to limit the selections for time with some general time frames in order to limit the resolution. I assumed this would make finding a common schedule a bit easier, but in retrospect it didn't seem too necessary. If this is good enough for you guys then great, if not let me know or feel free to recreate a schedule.